Remove old legacy search features/component tests
