--- es710_features/steps/fakeapi.go
@@ -1,50 +1,0 @@
-package steps
-
-import (
-	"fmt"
-	"io"
-	"net/http"
-	"testing"
-
-	"github.com/maxcnunes/httpfake"
-)
-
-// FakeAPI contains all the information for a fake component API
-type FakeAPI struct {
-	fakeHTTP                     *httpfake.HTTPFake
-	outboundRequests             []string
-	collectOutboundRequestBodies httpfake.CustomAssertor
-}
-
-// NewFakeAPI creates a new fake component API
-func NewFakeAPI(t testing.TB) *FakeAPI {
-	fa := &FakeAPI{
-		fakeHTTP: httpfake.New(httpfake.WithTesting(t)),
-	}
-
-	fa.collectOutboundRequestBodies = func(r *http.Request) error {
-		// inspect request
-		body, err := io.ReadAll(r.Body)
-		if err != nil {
-			return fmt.Errorf("error reading the outbound request body: %s", err.Error())
-		}
-		fa.outboundRequests = append(fa.outboundRequests, string(body))
-		return nil
-	}
-
-	return fa
-}
-
-func (f *FakeAPI) setJSONResponseForGetHealth(url string, statusCode int) {
-	f.fakeHTTP.NewHandler().Get(url).Reply(statusCode)
-}
-
-// Close closes the fake API
-func (f *FakeAPI) Close() {
-	f.fakeHTTP.Close()
-}
-
-// Reset resets the fake API
-func (f *FakeAPI) Reset() {
-	f.fakeHTTP.Reset()
-}
