--- es710_features/testdata/es_mulitple_search_release_results.json
@@ -1,213 +1,0 @@
-{
-  "took":11,
-  "responses":[
-    {
-      "took":8,
-      "timed_out":false,
-      "_shards":{
-        "total":5,
-        "successful":5,
-        "skipped":0,
-        "failed":0
-      },
-      "hits":{
-        "total":{
-          "value":16,
-          "relation":"eq"
-        },
-        "max_score":null,
-        "hits":[
-          {
-            "_index":"ons1646652430467",
-            "_type":"_doc",
-            "_id":"Estimating suicide among higher education students, England and Wales: Experimental Statistics",
-            "_score":0.2374177,
-            "_source":{
-              "type":"release",
-              "uri":"/releases/estimatingsuicideamonghighereducationstudentsenglandandwales",
-              "job_id":"",
-              "search_index":"",
-              "cdid":"",
-              "dataset_id":"",
-              "keywords":[],
-              "meta_description":"",
-              "release_date":"2018-06-25T08:30:00.000Z",
-              "summary":"Estimates of suicides among higher education students by sex, age and ethnicity. Analysis based on mortality records linked to Higher Education Statistics Agency (HESA) Student records. ",
-              "title":"Estimating suicide among higher education students, England and Wales: Experimental Statistics",
-              "date_changes":[
-                {
-                  "previous_date":"2018-06-01",
-                  "change_notice":"changed it"
-                }
-              ],
-              "topics":null,
-              "finalised":true,
-              "cancelled":false,
-              "published":false
-            },
-            "highlight":{
-              "summary":[
-                "Estimates of suicides among higher <em class=\"ons-highlight\">education</em> students by sex, age and ethnicity. Analysis based on mortality records linked to Higher <em class=\"ons-highlight\">Education</em> Statistics Agency (HESA) Student records. "
-              ],
-              "title":[
-                "Estimating suicide among higher <em class=\"ons-highlight\">education</em> students, England and <em class=\"ons-highlight\">Wales</em>: Experimental Statistics"
-              ]
-            },
-            "sort":[
-              0.2374177,
-              1529915400000
-            ]
-          },
-          {
-            "_index":"ons1646652430467",
-            "_type":"_doc",
-            "_id":"Young people not in education, employment or training (NEET), UK: August 2018",
-            "_score":0.04704748,
-            "_source":{
-              "uri":"/releases/youngpeoplenotineducationemploymentortrainingneetukaugust2018",
-              "type":"release",
-              "job_id":"",
-              "search_index":"",
-              "cdid":"",
-              "dataset_id":"",
-              "keywords":[],
-              "meta_description":"",
-              "title":"Young people not in education, employment or training (NEET), UK: August 2018",
-              "summary":"Estimates of young people (aged 16 to 24) who are not in education, employment or training, by age and sex.",
-              "release_date":"2018-08-23T08:30:00.000Z",
-              "topics":null,
-              "cancelled":false,
-              "finalised":false,
-              "published":false,
-              "provisional_date":"2018-08-23"
-            },
-            "highlight":{
-              "summary":[
-                "Estimates of young people (aged 16 to 24) who are not in <em class=\"ons-highlight\">education</em>, employment or training, by age and sex."
-              ],
-              "title":[
-                "Young people not in <em class=\"ons-highlight\">education</em>, employment or training (NEET), UK: August <em class=\"ons-highlight\">2018</em>"
-              ]
-            },
-            "sort":[
-              0.04704748,
-              1535013000000
-            ]
-          }
-        ]
-      },
-      "aggregations":{
-        "breakdown":{
-          "buckets":{
-            "provisional":{
-              "doc_count":1
-            },
-            "confirmed":{
-              "doc_count":0
-            },
-            "postponed":{
-              "doc_count":1
-            }
-          }
-        },
-        "census":{
-          "buckets":{
-            "census":{
-              "doc_count":0
-            }
-          }
-        }
-      }
-    },
-    {
-      "took":3,
-      "timed_out":false,
-      "_shards":{
-        "total":5,
-        "successful":5,
-        "skipped":0,
-        "failed":0
-      },
-      "hits":{
-        "total":{
-          "value":2979,
-          "relation":"eq"
-        },
-        "max_score":null,
-        "hits":[]
-      },
-      "aggregations":{
-        "release_types":{
-          "meta":{},
-          "buckets":{
-            "upcoming":{
-              "doc_count":25,
-              "breakdown":{
-                "buckets":{
-                  "provisional":{
-                    "doc_count":15
-                  },
-                  "postponed":{
-                    "doc_count":1
-                  },
-                  "confirmed":{
-                    "doc_count":9
-                  }
-                }
-              }
-            },
-            "outdated":{
-              "doc_count":473,
-              "breakdown":{
-                "buckets":{
-                  "provisional":{
-                    "doc_count":213
-                  },
-                  "postponed":{
-                    "doc_count":0
-                  },
-                  "confirmed":{
-                    "doc_count":260
-                  }
-                }
-              }
-            },
-            "published":{
-              "doc_count":2466,
-              "breakdown":{
-                "buckets":{
-                  "provisional":{
-                    "doc_count":5
-                  },
-                  "postponed":{
-                    "doc_count":0
-                  },
-                  "confirmed":{
-                    "doc_count":2461
-                  }
-                }
-              }
-            },
-            "cancelled":{
-              "doc_count":45,
-              "breakdown":{
-                "buckets":{
-                  "provisional":{
-                    "doc_count":18
-                  },
-                  "postponed":{
-                    "doc_count":0
-                  },
-                  "confirmed":{
-                    "doc_count":27
-                  }
-                }
-              }
-            }
-          }
-        }
-      },
-      "status":200
-    }
-  ]
-}
