--- es710_features/testdata/es_single_search_result.json
@@ -1,107 +1,0 @@
-{
-    "responses":[
-        {
-            "took":69,
-            "timed_out":false,
-            "_shards":{
-                "total":3,
-                "successful":3,
-                "failed":0
-            },
-            "hits":{
-                "total":1,
-                "max_score":null,
-                "hits":[
-                    {
-                        "_source":{
-                            "dataset_id": "/peoplepopulationandcommunity/housing/articles/housepricestatisticsforsmallareasinenglandandwales/2015-02-17",
-                            "keywords":[
-                                    "regional house prices",
-                                    "property prices",
-                                    "area with cheapest houses",
-                                    "area with most expensive houses"
-                            ],
-                            "release_date":"2015-02-17T00:00:00.000Z",
-                            "title":"House Price Statistics for Small Areas in England and Wales",
-                            "meta_description":"A brief overview of average (median) house prices using estimates from the sale prices of residential properties in England and Wales between 1995 and 2013.",
-                            "summary": "regional house prices",
-                            "type":"article",
-                            "uri":"/peoplepopulationandcommunity/housing/articles/housepricestatisticsforsmallareasinenglandandwales/2015-02-17"
-                        },
-                        "highlight":{
-                            "title":[
-                                "<em class=\"highlight\">House</em> Price Statistics for Small Areas in England and Wales"
-                            ],
-                            "edition":[
-                                "1995 <em class=\"highlight\">to</em> 2013"
-                            ],
-                            "summary":[
-                                "regional <em class=\"highlight\">house</em> prices"
-                            ],
-                            "metaDescription":[
-                                "A brief overview of average (median) <em class=\"highlight\">house</em> prices using estimates from the sale prices of residential properties in England and Wales between 1995 and 2013."
-                            ],
-                            "keywords":[
-                                "regional <em class=\"highlight\">house</em> prices",
-                                "area with cheapest <em class=\"highlight\">houses</em>",
-                                "area with most expensive <em class=\"highlight\">houses</em>"
-                            ],
-                            "datasetId":[
-                                "/peoplepopulationandcommunity/housing/articles/<em class=\"highlight\">house</em>pricestatisticsforsmallareasinenglandandwales/2015-02-17"
-                            ]
-                        },
-                        "sort":[
-                            51.435966,
-                            1424131200000
-                        ]
-                    }
-                ]
-            },
-            "suggest":{
-                "search_suggest":[
-                    {
-                        "text":"housing",
-                        "offset":0,
-                        "length":5,
-                        "options":[
-                            {
-                                "text":"home",
-                                "score":0.0849763
-                            },
-                            {
-                                "text":"tent",
-                                "score":0.0314159
-                            }
-                        ]
-                    }
-                ]
-            }
-        },
-        {
-            "took": 27,
-            "timed_out": false,
-            "_shards": {
-                "total": 3,
-                "successful": 3,
-                "failed": 0
-            },
-            "hits": {
-                "total": 2,
-                "max_score": 0.0,
-                "hits": []
-            },
-            "aggregations": {
-                "contentTypeCounts": {
-                    "doc_count_error_upper_bound": 0,
-                    "sum_other_doc_count": 0,
-                    "buckets": [
-                        {
-                            "key": "article",
-                            "doc_count": 1
-                        }
-                    ]
-                }
-            }
-        }
-    ]
-}
