--- es710_features/testdata/es_zero_search_release_results.json
@@ -1,126 +1,0 @@
-{
-  "responses":[
-    {
-      "took":8,
-      "timed_out":false,
-      "_shards":{
-        "total":3,
-        "successful":3,
-        "failed":0
-      },
-      "hits":{
-        "total":{
-          "value": 0
-        }
-      },
-      "aggregations":{
-        "breakdown":{
-          "buckets":{
-            "provisional":{
-              "doc_count":0
-            },
-            "confirmed":{
-              "doc_count":0
-            },
-            "postponed":{
-              "doc_count":0
-            }
-          }
-        },
-        "census":{
-          "buckets":{
-            "census":{
-              "doc_count":0
-            }
-          }
-        }
-      }
-    },
-    {
-      "took":3,
-      "timed_out":false,
-      "_shards":{
-        "total":3,
-        "successful":3,
-        "failed":0
-      },
-      "hits":{
-        "total":{
-          "value": 2979
-        },
-        "max_score":null,
-        "hits":[]
-      },
-      "aggregations":{
-        "release_types":{
-          "buckets":{
-            "upcoming":{
-              "doc_count":25,
-              "breakdown":{
-                "buckets":{
-                  "provisional":{
-                    "doc_count":15
-                  },
-                  "postponed":{
-                    "doc_count":1
-                  },
-                  "confirmed":{
-                    "doc_count":9
-                  }
-                }
-              }
-            },
-            "outdated":{
-              "doc_count":473,
-              "breakdown":{
-                "buckets":{
-                  "provisional":{
-                    "doc_count":213
-                  },
-                  "postponed":{
-                    "doc_count":0
-                  },
-                  "confirmed":{
-                    "doc_count":260
-                  }
-                }
-              }
-            },
-            "published":{
-              "doc_count":2466,
-              "breakdown":{
-                "buckets":{
-                  "provisional":{
-                    "doc_count":5
-                  },
-                  "postponed":{
-                    "doc_count":0
-                  },
-                  "confirmed":{
-                    "doc_count":2461
-                  }
-                }
-              }
-            },
-            "cancelled":{
-              "doc_count":45,
-              "breakdown":{
-                "buckets":{
-                  "provisional":{
-                    "doc_count":18
-                  },
-                  "postponed":{
-                    "doc_count":0
-                  },
-                  "confirmed":{
-                    "doc_count":27
-                  }
-                }
-              }
-            }
-          }
-        }
-      }
-    }
-  ]
-}
