--- es710_features/testdata/es_zero_search_results.json
@@ -1,42 +1,0 @@
-{
-    "responses": [
-        {
-            "took": 69,
-            "timed_out": false,
-            "_shards": {
-                "total": 3,
-                "successful": 3,
-                "failed": 0
-            },
-            "hits": {
-                "total": 0
-            },
-            "suggest": {
-                "search_suggest": [
-                ]
-            }
-        },
-        {
-            "took": 27,
-            "timed_out": false,
-            "_shards": {
-                "total": 3,
-                "successful": 3,
-                "failed": 0
-            },
-            "hits": {
-                "total": 0,
-                "max_score": 0.0,
-                "hits": []
-            },
-            "aggregations": {
-                "docCounts": {
-                    "doc_count_error_upper_bound": 0,
-                    "sum_other_doc_count": 0,
-                    "buckets": [
-                    ]
-                }
-            }
-        }
-    ]
-}
