--- es710_features/testdata/expected_multiple_search_results.json
@@ -1,60 +1,0 @@
-{
-    "es_710":true,
-    "count": 2,
-    "took": 96,
-    "topics": [],
-    "content_types": [
-        {
-            "type": "article",
-            "count": 1
-        },
-        {
-            "type": "product_page",
-            "count": 1
-        }
-    ],
-    "items": [
-        {
-            "dataset_id": "/peoplepopulationandcommunity/housing/articles/housepricestatisticsforsmallareasinenglandandwales/2015-02-17",
-            "highlight": {
-                "dataset_id": "/peoplepopulationandcommunity/housing/articles/<em class=\"highlight\">house</em>pricestatisticsforsmallareasinenglandandwales/2015-02-17",
-                "keywords": [
-                    "regional <em class=\"highlight\">house</em> prices",
-                    "property prices",
-                    "area with cheapest <em class=\"highlight\">houses</em>",
-                    "area with most expensive <em class=\"highlight\">houses</em>"
-                ],
-                "meta_description": "A brief overview of average (median) <em class=\"highlight\">house</em> prices using estimates from the sale prices of residential properties in England and Wales between 1995 and 2013.",
-                "summary": "regional <em class=\"highlight\">house</em> prices",
-                "title": "<em class=\"highlight\">House</em> Price Statistics for Small Areas in England and Wales"
-            },
-            "keywords": [
-                    "regional house prices",
-                    "property prices",
-                    "area with cheapest houses",
-                    "area with most expensive houses"
-            ],
-            "meta_description": "A brief overview of average (median) house prices using estimates from the sale prices of residential properties in England and Wales between 1995 and 2013.",
-            "release_date": "2015-02-17T00:00:00.000Z",
-            "summary": "regional house prices",
-            "title": "House Price Statistics for Small Areas in England and Wales",
-            "type": "article",
-            "uri": "/peoplepopulationandcommunity/housing/articles/housepricestatisticsforsmallareasinenglandandwales/2015-02-17"
-        },
-        {
-            "summary": "Data about property prices. These are an informal indicator of the confidence people feel in the UK economy. We collect the data from a sample of mortgage completions, so we do not include property sold to cash buyers.",
-            "highlight": {
-                "title": "<em class=\"highlight\">Housing</em>"
-            },
-            "title": "Housing",
-            "unit": "",
-            "source": "",
-            "type": "product_page",
-            "uri": "/peoplepopulationandcommunity/housing"
-        }
-    ],
-    "suggestions": [
-        "home",
-        "tent"
-    ]
-}
