--- es710_features/testdata/expected_single_search_result.json
@@ -1,45 +1,0 @@
-{
-    "es_710":true,
-    "count": 1,
-    "took": 96,
-    "topics": [],
-    "content_types": [
-        {
-            "type": "article",
-            "count": 1
-        }
-    ],
-    "items": [
-        {
-            "type": "article",
-            "dataset_id": "/peoplepopulationandcommunity/housing/articles/housepricestatisticsforsmallareasinenglandandwales/2015-02-17",
-            "keywords": [
-                "regional house prices",
-                "property prices",
-                "area with cheapest houses",
-                "area with most expensive houses"
-            ],
-            "meta_description": "A brief overview of average (median) house prices using estimates from the sale prices of residential properties in England and Wales between 1995 and 2013.",
-            "release_date": "2015-02-17T00:00:00.000Z",
-            "summary": "regional house prices",
-            "title": "House Price Statistics for Small Areas in England and Wales",
-            "uri": "/peoplepopulationandcommunity/housing/articles/housepricestatisticsforsmallareasinenglandandwales/2015-02-17",
-            "highlight": {
-                "dataset_id": "/peoplepopulationandcommunity/housing/articles/<em class=\"highlight\">house</em>pricestatisticsforsmallareasinenglandandwales/2015-02-17",
-                "keywords": [
-                    "regional <em class=\"highlight\">house</em> prices",
-                    "property prices",
-                    "area with cheapest <em class=\"highlight\">houses</em>",
-                    "area with most expensive <em class=\"highlight\">houses</em>"
-                ],
-                "meta_description": "A brief overview of average (median) <em class=\"highlight\">house</em> prices using estimates from the sale prices of residential properties in England and Wales between 1995 and 2013.",
-                "summary": "regional <em class=\"highlight\">house</em> prices",
-                "title": "<em class=\"highlight\">House</em> Price Statistics for Small Areas in England and Wales"
-            }
-        }
-    ],
-    "suggestions": [
-        "home",
-        "tent"
-    ]
-}
