--- es710_features/testdata/expected_zero_search_release_results.json
@@ -1,7 +1,0 @@
-{
-  "took":11,
-  "breakdown":{
-    "total":0
-  },
-  "releases":[]
-}
