--- es710_features/testdata/expected_zero_search_results.json
@@ -1,15 +1,0 @@
-{
-    "es_710":true,
-    "count": 0,
-    "took": 96,
-    "topics": [],
-    "content_types": [
-    ],
-    "items": [
-
-    ],
-    "additional_suggestions": [
-        "RPI",
-        "Consumers"
-    ]
-}
