+++ features/steps/fakeapi.go
@@ -35,8 +35,8 @@ func NewFakeAPI(t testing.TB) *FakeAPI {
 	return fa
 }
 
-func (f *FakeAPI) setJSONResponseForPost(url string, statusCode int, body []byte) {
-	f.fakeHTTP.NewHandler().Post(url).Reply(statusCode).Body(body)
+func (f *FakeAPI) setJSONResponseForGetHealth(url string, statusCode int) {
+	f.fakeHTTP.NewHandler().Get(url).Reply(statusCode)
 }
 
 // Close closes the fake API
