+++ features/steps/steps.go
@@ -12,92 +12,86 @@ import (
 )
 
 // RegisterSteps registers the specific steps needed to do component tests for the search api
-func (c *LegacyComponent) RegisterSteps(ctx *godog.ScenarioContext) {
+func (c *Component) RegisterSteps(ctx *godog.ScenarioContext) {
 	c.APIFeature.RegisterSteps(ctx)
-	ctx.Step(`^elasticsearch returns one item in search response$`, c.successfullyReturnSingleSearchResult)
+	ctx.Step(`^elasticsearch returns one item in search response$`, c.es7xSuccessfullyReturnSingleSearchResult)
 	ctx.Step(`^elasticsearch returns one item in search/release response$`, c.successfullyReturnSingleSearchReleaseResult)
-	ctx.Step(`^elasticsearch returns multiple items in search response$`, c.successfullyReturnMultipleSearchResults)
+	ctx.Step(`^the response body is the same as the json in "([^"]*)"$`, c.iShouldReceiveTheFollowingSearchResponsefromes7x)
+	ctx.Step(`^elasticsearch returns multiple items in search response$`, c.es7xSuccessfullyReturnMultipleSearchResults)
 	ctx.Step(`^elasticsearch returns multiple items in search/release response$`, c.successfullyReturnMultipleSearchReleaseResults)
-	ctx.Step(`^elasticsearch returns zero items in search response$`, c.successfullyReturnNoSearchResults)
+	ctx.Step(`^elasticsearch returns zero items in search response$`, c.es7xSuccessfullyReturnNoSearchResults)
 	ctx.Step(`^elasticsearch returns zero items in search/release response$`, c.successfullyReturnNoSearchReleaseResults)
-	ctx.Step(`^elasticsearch returns internal server error$`, c.failureInternalServerError)
-	ctx.Step(`^the response body is the same as the json in "([^"]*)"$`, c.iShouldReceiveTheFollowingSearchResponse)
+	ctx.Step(`^elasticsearch returns internal server error$`, c.es7xFailureInternalServerError)
 }
 
-func (c *LegacyComponent) successfullyReturnMultipleSearchResults() error {
-	body, err := os.ReadFile("./features/testdata/es_mulitple_search_results.json")
-	if err != nil {
-		return err
-	}
-
-	c.FakeElasticSearchAPI.setJSONResponseForPost("/elasticsearch/ons/_msearch", 200, body)
-
-	return nil
-}
-
-func (c *LegacyComponent) successfullyReturnMultipleSearchReleaseResults() error {
-	body, err := os.ReadFile("./features/testdata/es_mulitple_search_release_results.json")
-	if err != nil {
-		return err
-	}
-
-	c.FakeElasticSearchAPI.setJSONResponseForPost("/elasticsearch/ons/_msearch", 200, body)
-
-	return nil
-}
-
-func (c *LegacyComponent) successfullyReturnSingleSearchResult() error {
+func (c *Component) es7xSuccessfullyReturnSingleSearchResult() error {
 	body, err := os.ReadFile("./features/testdata/es_single_search_result.json")
 	if err != nil {
 		return err
 	}
 
-	c.FakeElasticSearchAPI.setJSONResponseForPost("/elasticsearch/ons/_msearch", 200, body)
+	c.FakeElasticSearchAPI.fakeHTTP.NewHandler().Get("/elasticsearch/_msearch").Reply(200).Body(body)
 
 	return nil
 }
 
-func (c *LegacyComponent) successfullyReturnSingleSearchReleaseResult() error {
+func (c *Component) successfullyReturnSingleSearchReleaseResult() error {
 	body, err := os.ReadFile("./features/testdata/es_single_search_release_result.json")
 	if err != nil {
 		return err
 	}
 
-	c.FakeElasticSearchAPI.setJSONResponseForPost("/elasticsearch/ons/_msearch", 200, body)
+	c.FakeElasticSearchAPI.fakeHTTP.NewHandler().Get("/elasticsearch/_msearch").Reply(200).Body(body)
 
 	return nil
 }
 
-func (c *LegacyComponent) successfullyReturnNoSearchResults() error {
+func (c *Component) es7xSuccessfullyReturnMultipleSearchResults() error {
+	body, err := os.ReadFile("./features/testdata/es_mulitple_search_results.json")
+	if err != nil {
+		return err
+	}
+
+	c.FakeElasticSearchAPI.fakeHTTP.NewHandler().Get("/elasticsearch/_msearch").Reply(200).Body(body)
+
+	return nil
+}
+
+func (c *Component) successfullyReturnMultipleSearchReleaseResults() error {
+	body, err := os.ReadFile("./features/testdata/es_mulitple_search_release_results.json")
+	if err != nil {
+		return err
+	}
+
+	c.FakeElasticSearchAPI.fakeHTTP.NewHandler().Get("/elasticsearch/_msearch").Reply(200).Body(body)
+
+	return nil
+}
+
+func (c *Component) es7xSuccessfullyReturnNoSearchResults() error {
 	body, err := os.ReadFile("./features/testdata/es_zero_search_results.json")
 	if err != nil {
 		return err
 	}
 
-	c.FakeElasticSearchAPI.setJSONResponseForPost("/elasticsearch/ons/_msearch", 200, body)
+	c.FakeElasticSearchAPI.fakeHTTP.NewHandler().Get("/elasticsearch/_msearch").Reply(200).Body(body)
 
 	return nil
 }
 
-func (c *LegacyComponent) successfullyReturnNoSearchReleaseResults() error {
+func (c *Component) successfullyReturnNoSearchReleaseResults() error {
 	body, err := os.ReadFile("./features/testdata/es_zero_search_release_results.json")
 	if err != nil {
 		return err
 	}
 
-	c.FakeElasticSearchAPI.setJSONResponseForPost("/elasticsearch/ons/_msearch", 200, body)
+	c.FakeElasticSearchAPI.fakeHTTP.NewHandler().Get("/elasticsearch/_msearch").Reply(200).Body(body)
 
 	return nil
 }
 
-func (c *LegacyComponent) failureInternalServerError() error {
-	c.FakeElasticSearchAPI.setJSONResponseForPost("/elasticsearch/ons/_msearch", 500, nil)
-
-	return nil
-}
-
-func (c *LegacyComponent) iShouldReceiveTheFollowingSearchResponse(expectedJSONFile string) error {
-	var searchResponse, expectedSearchResponse models.SearchResponseLegacy
+func (c *Component) iShouldReceiveTheFollowingSearchResponsefromes7x(expectedJSONFile string) error {
+	var searchResponse, expectedSearchResponse models.SearchResponse
 
 	responseBody, err := io.ReadAll(c.APIFeature.HttpResponse.Body)
 	if err != nil {
@@ -123,4 +117,10 @@ func (c *LegacyComponent) iShouldReceiveTheFollowingSearchResponse(expectedJSONF
 	}
 
 	return c.ErrorFeature.StepError()
+}
+
+func (c *Component) es7xFailureInternalServerError() error {
+	c.FakeElasticSearchAPI.fakeHTTP.NewHandler().Get("/elasticsearch/_msearch").Reply(500)
+
+	return nil
 }
