+++ features/testdata/es_mulitple_search_release_results.json
@@ -1,53 +1,55 @@
 {
+  "took":11,
   "responses":[
     {
       "took":8,
       "timed_out":false,
       "_shards":{
-        "total":3,
-        "successful":3,
+        "total":5,
+        "successful":5,
+        "skipped":0,
         "failed":0
       },
       "hits":{
-        "total":2,
-        "max_score":1.0,
+        "total":{
+          "value":16,
+          "relation":"eq"
+        },
+        "max_score":null,
         "hits":[
           {
             "_index":"ons1646652430467",
-            "_type":"release",
-            "_id":"/releases/estimatingsuicideamonghighereducationstudentsenglandandwales",
+            "_type":"_doc",
+            "_id":"Estimating suicide among higher education students, England and Wales: Experimental Statistics",
             "_score":0.2374177,
             "_source":{
-              "uri":"/releases/estimatingsuicideamonghighereducationstudentsenglandandwales",
-              "dateChanges": [{"previousDate": "2018-06-01", "changeNotice": "changed it"}],
               "type":"release",
-              "description":{
-                "finalised":true,
-                "title":"Estimating suicide among higher education students, England and Wales: Experimental Statistics",
-                "summary":"Estimates of suicides among higher education students by sex, age and ethnicity. Analysis based on mortality records linked to Higher Education Statistics Agency (HESA) Student records. ",
-                "nationalStatistic":false,
-                "contact":{
-                  "email":"mortality@ons.gov.uk",
-                  "name":"Sarah Caul",
-                  "telephone":"+44 (0)1633 456490"
-                },
-                "releaseDate":"2018-06-25T08:30:00.000Z",
-                "nextRelease":"",
-                "unit":"",
-                "preUnit":"",
-                "source":"",
-                "cancelled":false,
-                "cancellationNotice":[],
-                "published":false,
-                "provisionalDate":""
-              },
-              "searchBoost":[]
+              "uri":"/releases/estimatingsuicideamonghighereducationstudentsenglandandwales",
+              "job_id":"",
+              "search_index":"",
+              "cdid":"",
+              "dataset_id":"",
+              "keywords":[],
+              "meta_description":"",
+              "release_date":"2018-06-25T08:30:00.000Z",
+              "summary":"Estimates of suicides among higher education students by sex, age and ethnicity. Analysis based on mortality records linked to Higher Education Statistics Agency (HESA) Student records. ",
+              "title":"Estimating suicide among higher education students, England and Wales: Experimental Statistics",
+              "date_changes":[
+                {
+                  "previous_date":"2018-06-01",
+                  "change_notice":"changed it"
+                }
+              ],
+              "topics":null,
+              "finalised":true,
+              "cancelled":false,
+              "published":false
             },
             "highlight":{
-              "description.summary":[
+              "summary":[
                 "Estimates of suicides among higher <em class=\"ons-highlight\">education</em> students by sex, age and ethnicity. Analysis based on mortality records linked to Higher <em class=\"ons-highlight\">Education</em> Statistics Agency (HESA) Student records. "
               ],
-              "description.title":[
+              "title":[
                 "Estimating suicide among higher <em class=\"ons-highlight\">education</em> students, England and <em class=\"ons-highlight\">Wales</em>: Experimental Statistics"
               ]
             },
@@ -58,39 +60,32 @@
           },
           {
             "_index":"ons1646652430467",
-            "_type":"release",
-            "_id":"/releases/youngpeoplenotineducationemploymentortrainingneetukaugust2018",
+            "_type":"_doc",
+            "_id":"Young people not in education, employment or training (NEET), UK: August 2018",
             "_score":0.04704748,
             "_source":{
               "uri":"/releases/youngpeoplenotineducationemploymentortrainingneetukaugust2018",
               "type":"release",
-              "description":{
-                "finalised":false,
-                "title":"Young people not in education, employment or training (NEET), UK: August 2018",
-                "summary":"Estimates of young people (aged 16 to 24) who are not in education, employment or training, by age and sex.",
-                "nationalStatistic":true,
-                "contact":{
-                  "email":"yanitsa.petkova@ons.gov.uk",
-                  "name":"Yanitsa Petkova ",
-                  "telephone":"+44 (0)1633 651599 "
-                },
-                "releaseDate":"2018-08-23T08:30:00.000Z",
-                "nextRelease":"22 November 2018",
-                "unit":"",
-                "preUnit":"",
-                "source":"",
-                "cancelled":false,
-                "cancellationNotice":[],
-                "published":false,
-                "provisionalDate":"2018-08-23"
-              },
-              "searchBoost":[]
+              "job_id":"",
+              "search_index":"",
+              "cdid":"",
+              "dataset_id":"",
+              "keywords":[],
+              "meta_description":"",
+              "title":"Young people not in education, employment or training (NEET), UK: August 2018",
+              "summary":"Estimates of young people (aged 16 to 24) who are not in education, employment or training, by age and sex.",
+              "release_date":"2018-08-23T08:30:00.000Z",
+              "topics":null,
+              "cancelled":false,
+              "finalised":false,
+              "published":false,
+              "provisional_date":"2018-08-23"
             },
             "highlight":{
-              "description.summary":[
+              "summary":[
                 "Estimates of young people (aged 16 to 24) who are not in <em class=\"ons-highlight\">education</em>, employment or training, by age and sex."
               ],
-              "description.title":[
+              "title":[
                 "Young people not in <em class=\"ons-highlight\">education</em>, employment or training (NEET), UK: August <em class=\"ons-highlight\">2018</em>"
               ]
             },
@@ -128,17 +123,22 @@
       "took":3,
       "timed_out":false,
       "_shards":{
-        "total":3,
-        "successful":3,
+        "total":5,
+        "successful":5,
+        "skipped":0,
         "failed":0
       },
       "hits":{
-        "total":2979,
-        "max_score":0.0,
+        "total":{
+          "value":2979,
+          "relation":"eq"
+        },
+        "max_score":null,
         "hits":[]
       },
       "aggregations":{
         "release_types":{
+          "meta":{},
           "buckets":{
             "upcoming":{
               "doc_count":25,
@@ -206,7 +206,8 @@
             }
           }
         }
-      }
+      },
+      "status":200
     }
   ]
 }
