+++ features/testdata/es_single_search_result.json
@@ -13,61 +13,40 @@
                 "max_score":null,
                 "hits":[
                     {
-                        "_index":"ons1572283650465",
-                        "_type":"article",
-                        "_id":"/peoplepopulationandcommunity/housing/articles/housepricestatisticsforsmallareasinenglandandwales/2015-02-17",
-                        "_score":51.435966,
                         "_source":{
-                            "description":{
-                                "datasetId": "/peoplepopulationandcommunity/housing/articles/housepricestatisticsforsmallareasinenglandandwales/2015-02-17",
-                                "latestRelease":true,
-                                "unit":"number of people",
-                                "keywords":[
+                            "dataset_id": "/peoplepopulationandcommunity/housing/articles/housepricestatisticsforsmallareasinenglandandwales/2015-02-17",
+                            "keywords":[
                                     "regional house prices",
                                     "property prices",
                                     "area with cheapest houses",
                                     "area with most expensive houses"
-                                ],
-                                "releaseDate":"2015-02-17T00:00:00.000Z",
-                                "contact":{
-                                    "name":"Somebody's Name",
-                                    "telephone":"+44 (0)1329 123456",
-                                    "email":"somebody@ons.gov.uk"
-                                },
-                                "edition":"1995 to 2013",
-                                "source":"",
-                                "title":"House Price Statistics for Small Areas in England and Wales",
-                                "metaDescription":"A brief overview of average (median) house prices using estimates from the sale prices of residential properties in England and Wales between 1995 and 2013.",
-                                "nationalStatistic":false,
-                                "_abstract":"A brief overview of average (median) house prices using estimates from the sale prices of residential properties in England and Wales between 1995 and 2013. We look at the trends and features of average house prices paid over this period for 3 different geographies: local authorities, parliamentary constituencies and middle layer super output areas (MSOAs).",
-                                "preUnit":"",
-                                "summary": "regional house prices"
-                            },
-                            "searchBoost":[
-
                             ],
+                            "release_date":"2015-02-17T00:00:00.000Z",
+                            "title":"House Price Statistics for Small Areas in England and Wales",
+                            "meta_description":"A brief overview of average (median) house prices using estimates from the sale prices of residential properties in England and Wales between 1995 and 2013.",
+                            "summary": "regional house prices",
                             "type":"article",
                             "uri":"/peoplepopulationandcommunity/housing/articles/housepricestatisticsforsmallareasinenglandandwales/2015-02-17"
                         },
                         "highlight":{
-                            "description.title":[
+                            "title":[
                                 "<em class=\"highlight\">House</em> Price Statistics for Small Areas in England and Wales"
                             ],
-                            "description.edition":[
+                            "edition":[
                                 "1995 <em class=\"highlight\">to</em> 2013"
                             ],
-                            "description.summary":[
+                            "summary":[
                                 "regional <em class=\"highlight\">house</em> prices"
                             ],
-                            "description.metaDescription":[
+                            "metaDescription":[
                                 "A brief overview of average (median) <em class=\"highlight\">house</em> prices using estimates from the sale prices of residential properties in England and Wales between 1995 and 2013."
                             ],
-                            "description.keywords":[
+                            "keywords":[
                                 "regional <em class=\"highlight\">house</em> prices",
                                 "area with cheapest <em class=\"highlight\">houses</em>",
                                 "area with most expensive <em class=\"highlight\">houses</em>"
                             ],
-                            "description.datasetId":[
+                            "datasetId":[
                                 "/peoplepopulationandcommunity/housing/articles/<em class=\"highlight\">house</em>pricestatisticsforsmallareasinenglandandwales/2015-02-17"
                             ]
                         },
@@ -112,7 +91,7 @@
                 "hits": []
             },
             "aggregations": {
-                "docCounts": {
+                "contentTypeCounts": {
                     "doc_count_error_upper_bound": 0,
                     "sum_other_doc_count": 0,
                     "buckets": [
