+++ features/testdata/expected_multiple_search_results.json
@@ -1,6 +1,8 @@
 {
+    "es_710":true,
     "count": 2,
     "took": 96,
+    "topics": [],
     "content_types": [
         {
             "type": "article",
@@ -13,56 +15,40 @@
     ],
     "items": [
         {
-            "description": {
-                "contact": {
-                    "email": "somebody@ons.gov.uk",
-                    "name": "Somebody's Name",
-                    "telephone": "+44 (0)1329 123456"
-                },
-                "dataset_id": "/peoplepopulationandcommunity/housing/articles/housepricestatisticsforsmallareasinenglandandwales/2015-02-17",
-                "edition": "1995 to 2013",
-                "highlight": {
-                    "dataset_id": "/peoplepopulationandcommunity/housing/articles/<em class=\"highlight\">house</em>pricestatisticsforsmallareasinenglandandwales/2015-02-17",
-                    "edition": "1995 <em class=\"highlight\">to</em> 2013",
-                    "keywords": [
-                        "regional <em class=\"highlight\">house</em> prices",
-                        "property prices",
-                        "area with cheapest <em class=\"highlight\">houses</em>",
-                        "area with most expensive <em class=\"highlight\">houses</em>"
-                    ],
-                    "meta_description": "A brief overview of average (median) <em class=\"highlight\">house</em> prices using estimates from the sale prices of residential properties in England and Wales between 1995 and 2013.",
-                    "summary": "regional <em class=\"highlight\">house</em> prices",
-                    "title": "<em class=\"highlight\">House</em> Price Statistics for Small Areas in England and Wales"    
-                },
+            "dataset_id": "/peoplepopulationandcommunity/housing/articles/housepricestatisticsforsmallareasinenglandandwales/2015-02-17",
+            "highlight": {
+                "dataset_id": "/peoplepopulationandcommunity/housing/articles/<em class=\"highlight\">house</em>pricestatisticsforsmallareasinenglandandwales/2015-02-17",
                 "keywords": [
+                    "regional <em class=\"highlight\">house</em> prices",
+                    "property prices",
+                    "area with cheapest <em class=\"highlight\">houses</em>",
+                    "area with most expensive <em class=\"highlight\">houses</em>"
+                ],
+                "meta_description": "A brief overview of average (median) <em class=\"highlight\">house</em> prices using estimates from the sale prices of residential properties in England and Wales between 1995 and 2013.",
+                "summary": "regional <em class=\"highlight\">house</em> prices",
+                "title": "<em class=\"highlight\">House</em> Price Statistics for Small Areas in England and Wales"
+            },
+            "keywords": [
                     "regional house prices",
                     "property prices",
                     "area with cheapest houses",
                     "area with most expensive houses"
-                ],
-                "latest_release": true,
-                "meta_description": "A brief overview of average (median) house prices using estimates from the sale prices of residential properties in England and Wales between 1995 and 2013.",
-                "national_statistic": false,
-                "release_date": "2015-02-17T00:00:00.000Z",
-                "source": "",
-                "summary": "regional house prices",
-                "title": "House Price Statistics for Small Areas in England and Wales",
-                "unit": ""
-            },
+            ],
+            "meta_description": "A brief overview of average (median) house prices using estimates from the sale prices of residential properties in England and Wales between 1995 and 2013.",
+            "release_date": "2015-02-17T00:00:00.000Z",
+            "summary": "regional house prices",
+            "title": "House Price Statistics for Small Areas in England and Wales",
             "type": "article",
             "uri": "/peoplepopulationandcommunity/housing/articles/housepricestatisticsforsmallareasinenglandandwales/2015-02-17"
         },
         {
-            "description": {
-                "highlight": {
-                    "title": "<em class=\"highlight\">Housing</em>"
-                },
-                "summary": "Data about property prices. These are an informal indicator of the confidence people feel in the UK economy. We collect the data from a sample of mortgage completions, so we do not include property sold to cash buyers.",
-                "title": "Housing",
-                "unit": "",
-                "source": "",
-                "pre_unit": ""
+            "summary": "Data about property prices. These are an informal indicator of the confidence people feel in the UK economy. We collect the data from a sample of mortgage completions, so we do not include property sold to cash buyers.",
+            "highlight": {
+                "title": "<em class=\"highlight\">Housing</em>"
             },
+            "title": "Housing",
+            "unit": "",
+            "source": "",
             "type": "product_page",
             "uri": "/peoplepopulationandcommunity/housing"
         }
