--- features/testdata/expected_search_results_without_highlight.json
@@ -1,58 +1,0 @@
-{
-    "count": 2,
-    "took": 96,
-    "content_types": [
-        {
-            "type": "article",
-            "count": 1
-        },
-        {
-            "type": "product_page",
-            "count": 1
-        }
-    ],
-    "items": [
-        {
-            "description": {
-                "contact": {
-                    "email": "somebody@ons.gov.uk",
-                    "name": "Somebody's Name",
-                    "telephone": "+44 (0)1329 123456"
-                },
-                "dataset_id": "/peoplepopulationandcommunity/housing/articles/housepricestatisticsforsmallareasinenglandandwales/2015-02-17",
-                "edition": "1995 to 2013",
-                "keywords": [
-                    "regional house prices",
-                    "property prices",
-                    "area with cheapest houses",
-                    "area with most expensive houses"
-                ],
-                "latest_release": true,
-                "meta_description": "A brief overview of average (median) house prices using estimates from the sale prices of residential properties in England and Wales between 1995 and 2013.",
-                "national_statistic": false,
-                "release_date": "2015-02-17T00:00:00.000Z",
-                "source": "",
-                "summary": "regional house prices",
-                "title": "House Price Statistics for Small Areas in England and Wales",
-                "unit": ""
-            },
-            "type": "article",
-            "uri": "/peoplepopulationandcommunity/housing/articles/housepricestatisticsforsmallareasinenglandandwales/2015-02-17"
-        },
-        {
-            "description": {
-                "summary": "Data about property prices. These are an informal indicator of the confidence people feel in the UK economy. We collect the data from a sample of mortgage completions, so we do not include property sold to cash buyers.",
-                "title": "Housing",
-                "unit": "",
-                "source": "",
-                "pre_unit": ""
-            },
-            "type": "product_page",
-            "uri": "/peoplepopulationandcommunity/housing"
-        }
-    ],
-    "suggestions": [
-        "home",
-        "tent"
-    ]
-}
