+++ features/testdata/expected_single_search_release_result.json
@@ -18,14 +18,7 @@
         "cancelled":false,
         "finalised":true,
         "postponed":true,
-        "census":false,
-        "national_statistic":false,
-        "latest_release":false,
-        "contact":{
-          "name":"Sarah Caul",
-          "telephone":"+44 (0)1633 456490",
-          "email":"mortality@ons.gov.uk"
-        }
+        "census":false
       },
       "highlight":{
         "summary":"Estimates of suicides among higher \u003cem class=\"ons-highlight\"\u003eeducation\u003c/em\u003e students by sex, age and ethnicity. Analysis based on mortality records linked to Higher \u003cem class=\"ons-highlight\"\u003eEducation\u003c/em\u003e Statistics Agency (HESA) Student records. ",
