+++ features/testdata/expected_single_search_release_result_nohighlight.json
@@ -18,14 +18,7 @@
         "cancelled":false,
         "finalised":true,
         "postponed":true,
-        "census":false,
-        "national_statistic":false,
-        "latest_release":false,
-        "contact":{
-          "name":"Sarah Caul",
-          "telephone":"+44 (0)1633 456490",
-          "email":"mortality@ons.gov.uk"
-        }
+        "census":false
       }
     }
   ]
