+++ features/testdata/expected_zero_search_results.json
@@ -1,6 +1,8 @@
 {
+    "es_710":true,
     "count": 0,
     "took": 96,
+    "topics": [],
     "content_types": [
     ],
     "items": [
